+++ test/seating_test.rb
@@ -2,9 +2,21 @@ require_relative "test_helper"
 
 class SeatingTest < Minitest::Test
   def test_works
-    connections = []
-    tables = []
+    connections = [
+      {people: ["A", "B", "C"], strength: 2},
+      {people: ["B", "C", "D", "E"], strength: 1}
+    ]
+    tables = [3, 2]
     seating = ORTools::Seating.new(connections: connections, tables: tables)
-    p seating.assignments
+    assert_equal ["A", "B", "C", "D", "E"], seating.people
+    assert_equal({"A" => 2, "B" => 3, "D" => 1, "E" => 1}, seating.connections_for("C"))
+    expected = [
+      {person: "A", table: 0},
+      {person: "B", table: 0},
+      {person: "C", table: 0},
+      {person: "D", table: 1},
+      {person: "E", table: 1}
+    ]
+    assert_equal expected, seating.assignments
   end
 end